+++ src/main/java/net/geforcemods/securitycraft/tileentity/TileEntityInventoryScanner.java
@@ -52,7 +52,7 @@ public class TileEntityInventoryScanner extends TileEntityDisguisable implements
 
 	@Override
 	public void onOwnerChanged(IBlockState state, World world, BlockPos pos, EntityPlayer player) {
-		TileEntityInventoryScanner connectedScanner = BlockInventoryScanner.getConnectedInventoryScanner(world, pos);
+		TileEntityInventoryScanner connectedScanner = BlockInventoryScanner.getConnectedInventoryScanner(world, pos, be -> be.setOwner(getOwner().getUUID(), getOwner().getName()));
 
 		if (connectedScanner != null) {
 			connectedScanner.setOwner(getOwner().getUUID(), getOwner().getName());